ci: add multi-layer dependency caching
Split build.ts into pack:build and pack:install scripts so CI can
cache dist/node_modules between deploys. Add npm download cache
layer via setup-node for faster cache-miss installs.
